Added DiHelper. #13
